Move klog setup in a dedicated package
